Add validators for settings
